Fix suggestions
#SQA-806
